# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Service Instances",
            "description": "Verwaltung von Eureka-Service-Instances: Abfrage konfigurierter und aktiv registrierter Instanzen sowie Aktualisierung von Verbindungsparametern (Host, IP, Port) mit automatischer De-Registrierung und Neuregistrierung bei Eureka."
        }
    ],
    "paths": {
        "/instances/update": {
            "put": {
                "tags": [
                    "Service Instances"
                ],
                "summary": "Verbindungsparameter einer Instanz aktualisieren",
                "description": "Aktualisiert Host, IP-Adresse und/oder Ports einer laufenden Service-Instance. Der Lifecycle-Manager meldet die Instanz zunächst von Eureka ab, übernimmt die neuen Parameter und registriert die Instanz anschließend neu. Der Vorgang ist idempotent, sofern dieselben Parameter erneut übergeben werden.",
                "operationId": "updateInstance",
                "requestBody": {
                    "description": "Neue Verbindungsparameter der zu aktualisierenden Service-Instance. Pflichtfeld: serviceName (Identifikation). Alle übrigen Felder überschreiben die bisherige Konfiguration.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInstanceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Instanz erfolgreich aktualisiert und neu bei Eureka registriert.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Instanz 'my-service' erfolgreich aktualisiert und neu gestartet."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Keine konfigurierte Instance mit dem angegebenen serviceName gefunden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "ServiceInstance 'unknown-service' nicht gefunden."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/running": {
            "get": {
                "tags": [
                    "Service Instances"
                ],
                "summary": "Aktiv registrierte Instanzen abrufen",
                "description": "Gibt alle Service-Instances zurück, die aktuell aktiv bei Eureka registriert sind und Heartbeats senden (Status UP). Instanzen, die konfiguriert, aber noch nicht gestartet oder bereits abgemeldet wurden, erscheinen hier nicht.",
                "operationId": "getRunningInstances",
                "responses": {
                    "200": {
                        "description": "Aktiv registrierte Instanzen erfolgreich zurückgegeben (leere Liste möglich).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceInstance"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/instances/configured": {
            "get": {
                "tags": [
                    "Service Instances"
                ],
                "summary": "Konfigurierte Instanzen abrufen",
                "description": "Gibt alle Service-Instances zurück, die statisch in der Anwendungskonfiguration (z. B. application.yml) definiert sind – unabhängig davon, ob sie aktuell bei Eureka registriert sind.",
                "operationId": "getConfiguredInstances",
                "responses": {
                    "200": {
                        "description": "Konfigurierte Instanzen erfolgreich zurückgegeben (leere Liste möglich).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceInstance"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateInstanceRequest": {
                "type": "object",
                "description": "Request-Body zum Aktualisieren der Verbindungsparameter einer Service-Instance.",
                "properties": {
                    "serviceName": {
                        "type": "string",
                        "description": "Name der zu aktualisierenden Service-Instance (Pflichtfeld, dient als Identifikation).",
                        "example": "payment-service"
                    },
                    "newHostName": {
                        "type": "string",
                        "description": "Neuer Hostname, unter dem die Instance erreichbar ist.",
                        "example": "payment-host-02"
                    },
                    "newIpAddress": {
                        "type": "string",
                        "description": "Neue IP-Adresse der Instance.",
                        "example": "192.168.1.99"
                    },
                    "httpPort": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Neuer HTTP-Port der Instance.",
                        "example": 8081
                    },
                    "securePort": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Neuer HTTPS/Secure-Port der Instance (0 = deaktiviert).",
                        "example": 8444
                    },
                    "sslPreferred": {
                        "type": "boolean",
                        "description": "Gibt an, ob Clients den Secure-Port bevorzugen sollen.",
                        "example": false
                    }
                },
                "required": [
                    "serviceName"
                ]
            },
            "ServiceInstance": {
                "type": "object",
                "description": "Repräsentiert eine bei Eureka registrierbare oder registrierte Service-Instance.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Interne ID der Instance.",
                        "example": 1
                    },
                    "serviceName": {
                        "type": "string",
                        "description": "Eindeutiger Name des Services, unter dem er sich bei Eureka registriert.",
                        "example": "payment-service"
                    },
                    "hostName": {
                        "type": "string",
                        "description": "Hostname der Instance, den Eureka an Clients kommuniziert.",
                        "example": "payment-host-01"
                    },
                    "httpPort": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP-Port, auf dem der Service erreichbar ist.",
                        "example": 8080
                    },
                    "securePort": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTPS/Secure-Port des Services (0 = deaktiviert).",
                        "example": 8443
                    },
                    "ipAddr": {
                        "type": "string",
                        "description": "IP-Adresse der Instance.",
                        "example": "192.168.1.42"
                    },
                    "dataCenterInfoName": {
                        "type": "string",
                        "description": "Name des Data-Center-Providers (z. B. 'MyOwn' oder 'Amazon').",
                        "example": "MyOwn"
                    },
                    "status": {
                        "type": "string",
                        "description": "Aktueller Eureka-Status der Instance (UP, DOWN, STARTING, OUT_OF_SERVICE).",
                        "example": "UP"
                    },
                    "sslPreferred": {
                        "type": "boolean",
                        "description": "Gibt an, ob Clients den Secure-Port bevorzugen sollen.",
                        "example": false
                    }
                }
            }
        }
    }
}